changed output of regfile name to data_readRegA/B
this is a backup to my processor code uploaded to the 350 github repository.  i submitted my final working processor this afternoon, and then realized this evening that i forgot to change the output of my regfile name to match the skeleton.  when i did this and began checking it, crazy things started happening with my waveforms, so i decided to just make a note about it in my README in what i already submitted, rather than risk messing it all up.  this simply serves as a back up repository on the off chance that something was corrupted before i changed the output names.